Analyze a categorical color palette

Analyze a categorical color palette with respect to the differences between the colors in the palette.

Usage

analyze_palette(
  palette,
  cvd = c("protan", "deutan", "tritan"),
  cvd_severity = 1,
  bg = NULL,
  metric = c("din99d", "ciede2000", "cie76")
)

Arguments

palette

Either a matrix of RGB values (with values between 0 and 1), a data frame with RGB values, or a character vector of hex colors.

cvd

Color vision deficiency adaptation. Use cvd_severity to set the severity of color vision deficiency to adapt to. Permissible values are "protan", "deutan", and "tritan".

cvd_severity

Severity of color vision deficiency to adapt to. Can take any value from 0, for normal vision (the default), and 1, for dichromatic vision.

bg

Backgrouind color to consider (but not include) when generating the palette. This is useful to avoid colros that are too close to the background/canvas color. If NULL (the default), the background color is not considered at all. Any color that is convertable via col2rgb is acceptable, including hex colors.

metric

The color metric to use for the color distance matrix.

Value

A list of lists, one for each type of color vision deficiency plus normal vision. Each list contains difference_matrix, min_distances, and bg_min_distance

See also

Examples

pal <- qualpal(5)
analyze_palette(pal$hex, cvd = "protan", cvd_severity = 0.5)
#> $normal
#> $normal$difference_matrix
#>          #6d6cc9  #73c6c0  #c8cb70  #cb7469
#> #73c6c0 22.79624                           
#> #c8cb70 29.28845 19.20223                  
#> #cb7469 23.10335 25.67505 20.42218         
#> #e2cbe5 19.34558 19.26497 21.05212 18.54570
#> 
#> $normal$min_distances
#>  #6d6cc9  #73c6c0  #c8cb70  #cb7469  #e2cbe5 
#> 19.34558 19.20223 19.20223 18.54570 18.54570 
#> 
#> $normal$bg_min_distance
#> [1] NaN
#> 
#> 
#> $protan
#> $protan$difference_matrix
#>          #6d6cc9  #73c6c0  #c8cb70  #cb7469
#> #73c6c0 19.12270                           
#> #c8cb70 29.02227 18.74962                  
#> #cb7469 22.09733 18.62002 16.44912         
#> #e2cbe5 18.67437 11.58798 20.50475 19.14391
#> 
#> $protan$min_distances
#>  #6d6cc9  #73c6c0  #c8cb70  #cb7469  #e2cbe5 
#> 18.67437 11.58798 16.44912 16.44912 11.58798 
#> 
#> $protan$bg_min_distance
#> [1] NaN
#> 
#>