Generate qualitative color palettes

Given a collection of colors, qualpal() algorithmically tries to select to n most distinct colors from the provided input colors, optionally taking color vision deficiency into account.

Usage

qualpal(
  n,
  colorspace = list(h = c(0, 360), s = c(0.2, 0.5), l = c(0.6, 0.85)),
  cvd = c(protan = 0, deutan = 0, tritan = 0),
  cvd_severity,
  bg = NULL,
  metric = c("din99d", "ciede2000", "cie76"),
  extend = NULL
)

Arguments

n

The number of colors to generate.

colorspace

A color space to generate colors from. Can be any of the following:

A list that describes a color space in either HSL or LCHab color space. In the first case (HSL), the list must contain the following named vectors, each of length two, giving a range for each item.

h

Hue, in range from -360 to 360

s

Saturation, in the range from 0 to 1

l

Lightness, in the range from 0 to 1

In the second case (LCHab), the list must contain the following named vectors, each of length two, giving a range for each item.

h

Hue, in range from -360 to 360

c

Chroma, in the range from 0 to infinity

l

Lightness, in the range from 0 to 100

In these cases, qualpal() will generate

A character vector of length one in the form of "Source:Palette", where Domain is the name of a source that provides a color palette, and Palette is the name of a color palette from that source. See list_palettes() for available palettes.

cvd

Color vision deficiency adaptation. This must be a named vector with names protan, deutan, and tritan and values between 0 and 1, where 0 means no adaptation and 1 means full adaptation.

cvd_severity

DEPRECATED. Use a named cvd vector instead, e.g. c(protan = 0.5, deutan = 0.2, tritan = 0).

bg

Backgrouind color to consider (but not include) when generating the palette. This is useful to avoid colros that are too close to the background/canvas color. If NULL (the default), the background color is not considered at all. Any color that is convertable via col2rgb is acceptable, including hex colors.

metric

The color metric to use for the color distance matrix.

extend

A palette of colors to use as a fixed set of initial colors in the palette, which can be either a matrix or data frame of RGB values (with values between 0 and 1) or a character vector of hex colors (or any other format that's acceptable in grDevices::col2rgb()).

Value

A list of class qualpal with the following components.

HSL

A matrix of the colors in the HSL color space.

RGB

A matrix of the colors in the sRGB color space.

hex

A character vector of the colors in hex notation.

de_DIN99d

A distance matrix of color differences according to delta E DIN99d.

DIN99d

A color difference matrix of the colors corrosponding to the metric used.

de_DIN99d

The minimum pairwise DIN99d color difference for each color in the palette.

hex

A character vector of the colors in hex notation.

min_de_DIN99d

The minimum pairwise DIN99d color difference among all colors in the palette.

Details

In the default metric setting, the colors that the user provides are projected into the DIN99d color space, which is approximately perceptually uniform, i.e. color difference is proportional to the euclidean distance between two colors. A distance matrix is computed and, as an additional step, is transformed using power transformations discovered by Huang 2015 in order to fine tune differences.

qualpal then searches the distance matrix for the most distinct colors; it does this iteratively by first selecting a random set of colors and then iterates over each color, putting colors back into the total set and replaces it with a new color until it has gone through the whole range without changing any of the colors.

Optionally, qualpal can adapt palettes to cater to color vision deficiency (CVD). This is accomplished by taking the colors provided by the user and transforming them to colors that someone with CVD would see, that is, simulating CVD qualpal then chooses colors from these new colors.

See also

Examples

# Generate 3 distinct colors from the default color space
qualpal(3)
#> ---------------------------------------- 
#> Colors in the HSL color space 
#> 
#>         Hue Saturation Lightness
#> #b9d179  76       0.49      0.65
#> #c96c74 355       0.47      0.61
#> #6f99c5 210       0.43      0.61
#> 
#>  ---------------------------------------- 
#> DIN99d color difference distance matrix 
#> 
#>         #b9d179 #c96c74
#> #c96c74      24        
#> #6f99c5      24      23

# Provide a custom color space
qualpal(n = 3, list(h = c(35, 360), s = c(0.5, 0.7), l = c(0, 0.45)))
#> ---------------------------------------- 
#> Colors in the HSL color space 
#> 
#>         Hue Saturation Lightness
#> #9eba25  71       0.66     0.440
#> #b32eb6 298       0.60     0.449
#> #062010 143       0.68     0.076
#> 
#>  ---------------------------------------- 
#> DIN99d color difference distance matrix 
#> 
#>         #9eba25 #b32eb6
#> #b32eb6      32        
#> #062010      29      29

qualpal(3, "ColorBrewer:Set2")
#> ---------------------------------------- 
#> Colors in the HSL color space 
#> 
#>         Hue Saturation Lightness
#> #8da0cb 222       0.37      0.67
#> #a6d854  83       0.63      0.59
#> #fc8d62  17       0.96      0.69
#> 
#>  ---------------------------------------- 
#> DIN99d color difference distance matrix 
#> 
#>         #8da0cb #a6d854
#> #a6d854      26        
#> #fc8d62      24      22

# Adapt palette to deuteranopia
qualpal(5, "ColorBrewer:Dark2", cvd = c(deutan = 1))
#> ---------------------------------------- 
#> Colors in the HSL color space 
#> 
#>         Hue Saturation Lightness
#> #e6ab02  44       0.98      0.45
#> #a6761d  39       0.70      0.38
#> #666666   0       0.00      0.40
#> #7570b3 244       0.31      0.57
#> #e7298a 329       0.80      0.53
#> 
#>  ---------------------------------------- 
#> DIN99d color difference distance matrix 
#> 
#>         #e6ab02 #a6761d #666666 #7570b3
#> #a6761d      12                        
#> #666666      23      18                
#> #7570b3      29      26      15        
#> #e7298a      26      22      21      20

# Adapt palette to protanomaly with severity 0.4
qualpal(8, cvd = c(protan = 0.4))
#> ---------------------------------------- 
#> Colors in the HSL color space 
#> 
#>           Hue Saturation Lightness
#> #cb7469   6.7       0.49      0.60
#> #6d6cc9 240.8       0.47      0.61
#> #c8cb70  61.5       0.47      0.62
#> #d074a1 330.5       0.50      0.64
#> #7cb697 147.7       0.29      0.60
#> #eacbc5   9.8       0.48      0.85
#> #e8bbe5 304.1       0.49      0.82
#> #81cbd5 187.0       0.50      0.67
#> 
#>  ---------------------------------------- 
#> DIN99d color difference distance matrix 
#> 
#>         #cb7469 #6d6cc9 #c8cb70 #d074a1 #7cb697 #eacbc5 #e8bbe5
#> #6d6cc9      23                                                
#> #c8cb70      20      29                                        
#> #d074a1      12      18      25                                
#> #7cb697      23      24      15      25                        
#> #eacbc5      15      23      16      16      19                
#> #e8bbe5      18      19      23      13      22      12        
#> #81cbd5      26      21      21      25      11      19      20

# Generate and extend a palette with 3 colors
pal <- qualpal(3)
qualpal(5, extend = pal$hex)
#> ---------------------------------------- 
#> Colors in the HSL color space 
#> 
#>         Hue Saturation Lightness
#> #b9d179  76       0.49      0.65
#> #c96c74 355       0.46      0.61
#> #6f99c5 211       0.43      0.60
#> #e0d0d9 326       0.20      0.85
#> #ab6dc6 282       0.44      0.60
#> 
#>  ---------------------------------------- 
#> DIN99d color difference distance matrix 
#> 
#>         #b9d179 #c96c74 #6f99c5 #e0d0d9
#> #c96c74      24                        
#> #6f99c5      24      23                
#> #e0d0d9      19      18      17        
#> #ab6dc6      28      17      17      19

if (FALSE) { # \dontrun{
# The range of hue cannot exceed 360
qualpal(3, list(h = c(-20, 360), s = c(0.5, 0.7), l = c(0, 0.45)))
} # }